supervisor/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyserial==3.5
pyserial-asyncio==0.6        # <— this provides import serial_asyncio
pandas==2.2.2
pyarrow==17.0.0
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2                # <— use real YAML instead of the hack
curl